# api/src/lib/openapi/base.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenZenith API",
    "version": "0.9.2",
    "description": "Free, fast, global elevation and geospatial API. Query any point on Earth for elevation data from NASA SRTM 30m, track flights, monitor weather, explore OpenStreetMap, and more. No API key required.",
    "contact": {
      "name": "OpenZenith",
      "url": "https://github.com/aliasfoxkde/OpenZenith"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment (derived from request)"
    }
  ],
  "paths": {
...
    "/api/hurricanes": {
      "get": {
        "summary": "Hurricane track data from IBTrACS",
        "description": "Returns tropical cyclone track data as GeoJSON from the IBTrACS dataset.",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "tags": ["Data"]
      }
    },
...
    "/api/elevation/batch": {
      "post": {
        "summary": "Batch elevation lookup",
        "description": "Returns elevation for up to 2000 geographic points in a single request. Accepts the same `interpolation`, `units` and `datum` query parameters as `/api/elevation`; each result carries the raw sample in `elevation_m` and the response carries a `metadata` block.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["points"],
                "properties": {
                  "points": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["lat", "lon"],
                      "properties": {
                        "lat": {
                          "type": "number"
                        },
                        "lon": {
                          "type": "number"
                        },
                        "id": {
                          "type": "string"
                        }
                      }
                    },
                    "maxItems": 2000
                  }
                }
              }
            }
          }
        },
        "tags": ["Elevation"],
        "parameters": [
          {
            "name": "interpolation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["nearest", "bilinear"],
              "default": "bilinear"
            },
            "description": "Grid sampling: `bilinear` blends the 4 neighbouring pixels (default), `nearest` returns the single containing pixel. Sources without a 4-pixel neighbourhood (merged chunks, GEBCO) always sample nearest and report it in `metadata.interpolation`."
          },
          {
            "name": "units",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["meters", "feet"],
              "default": "meters"
            },
            "description": "Unit of `elevation`. The raw EGM96 orthometric sample stays in metres in `metadata.elevation_m`."
          },
          {
            "name": "datum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["egm96", "ellipsoid"],
              "default": "egm96"
            },
            "description": "Vertical datum. `egm96` (default) returns the orthometric height SRTM already carries; `ellipsoid` returns h = H + N using the bundled 30' EGM96 grid and reports N in `metadata.geoid_undulation_m`."
          }
        ],
        "responses": {
          "200": {
            "description": "Batch elevation data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "lat": {
                            "type": "number"
                          },
                          "lon": {
                            "type": "number"
                          },
                          "elevation": {
                            "type": ["number", "null"],
                            "description": "In the requested `units` and `datum`"
                          },
                          "elevation_m": {
                            "type": ["number", "null"],
                            "description": "Raw EGM96 orthometric sample in metres"
                          }
                        }
                      }
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "resolution_m": {
                          "type": "number",
                          "example": 30
                        },
                        "vertical_datum": {
                          "type": "string",
                          "enum": ["egm96", "ellipsoid"]
                        },
                        "interpolation": {
                          "type": "string",
                          "enum": ["nearest", "bilinear"]
                        },
                        "units": {
                          "type": "string",
                          "enum": ["meters", "feet"]
                        },
                        "source": {
                          "type": "string",
                          "example": "huggingface"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}